requirements.txt
absl-py==2.0.0
# apturl==0.5.2
astunparse==1.6.3
bcrypt==3.2.0
blessings==1.7
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.0
docutils==0.17.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.4.0
future==0.18.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg===1.16.0
grpcio==1.59.0
h5py==3.10.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
-e git+https://github.com/median-research-group/LibMTL@f10f7c9ffb72138a4ffae150330fb653da3b7456#egg=LibMTL
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.2.1
netifaces==0.11.0
networkx==3.0rc1
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
POT==0.9.1
protobuf==4.24.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytorch-triton
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
roman==3.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
sympy==1.11.1
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
trash-cli==0.17.1.14
typing_extensions
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
usb-creator==0.3.7
tywadllib
Werkzeug==3.0.0
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
